refactoring changes

+++ src/portfft/dispatcher/workgroup_dispatcher.hpp
@@ -391,7 +391,8 @@ struct committed_descriptor_impl<Scalar, Domain>::calculate_twiddles_struct::inn
         sycl::aligned_alloc_device<Scalar>(alignof(sycl::vec<Scalar, PORTFFT_VEC_LOAD_BYTES / sizeof(Scalar)>),
                                            static_cast<std::size_t>(res_size), desc.queue);
     if (!res) {
-      throw internal_error("Could not allocate usm memory of size: ", res_size * sizeof(Scalar), " bytes");
+      throw internal_error(
+          "Could not allocate usm memory of size: ", static_cast<std::size_t>(res_size) * sizeof(Scalar), " bytes");
     }
     desc.queue.submit([&](sycl::handler& cgh) {
       PORTFFT_LOG_TRACE(
